Fixed id_toa() related memory leaks.
